Rename `strings` module to `resources`
