+++ resources/build.gradle.kts
