+++ resources/src/commonMain/kotlin/dev/sasikanth/rss/reader/resources/strings/EnTwineStrings.kt
@@ -1,4 +1,4 @@
-package dev.sasikanth.rss.reader.resources
+package dev.sasikanth.rss.reader.resources.strings
 
 import cafe.adriel.lyricist.LyricistStrings
